% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data}
\usage{
test_data
}
\description{
A simulated data frame for illustration. In the test data, \eqn{X_1} and \eqn{X_2} are continuous variables, \eqn{X_3}
and \eqn{X_4} are binary variables, \eqn{Y} is the continuous outcome, and \eqn{Z} is the dichotomous treatment indicator.
}
